refactor: Lottos -> LottoBundle 클래스명 수정 및 변수, 메서드명 수정
